Avoid changing project dependencies for now
