--- lib/scala/logging-system2slf4j/src/main/java/module-info.java
@@ -1,8 +1,0 @@
-import org.enso.logging.system2slf4j.SystemLoggerViaSlf4j;
-
-module org.enso.logging.system2slf4j {
-  requires org.slf4j;
-
-  provides java.lang.System.LoggerFinder with
-      SystemLoggerViaSlf4j;
-}
